fix: rejig into strategy/factory impl
