+++ internal/cmdutils/cmdutils.go
@@ -20,15 +20,16 @@ const (
 	BOLD_TERMINAL    string = "\x1b[1m%s"
 )
 
-func DisplayTaskSelection(conf *config.Config) (taskOrPipelineSelected string, err error) {
+func DisplayTaskSelection(conf *config.Config, showPipelineOnly bool) (taskOrPipelineSelected string, err error) {
 	optionMap := []huh.Option[string]{}
 
 	for pipeline := range conf.Pipelines {
 		optionMap = append(optionMap, huh.NewOption(fmt.Sprintf("%s - %s", pipeline, fmt.Sprintf(GREY_TERMINAL, "pipeline")), pipeline))
 	}
-
-	for _, task := range conf.Tasks {
-		optionMap = append(optionMap, huh.NewOption(fmt.Sprintf("%s - %s", task.Name, fmt.Sprintf(GREY_TERMINAL, task.Description)), task.Name)) // fmt.Sprintf("Task: %s", task.Name)
+	if !showPipelineOnly {
+		for _, task := range conf.Tasks {
+			optionMap = append(optionMap, huh.NewOption(fmt.Sprintf("%s - %s", task.Name, fmt.Sprintf(GREY_TERMINAL, task.Description)), task.Name)) // fmt.Sprintf("Task: %s", task.Name)
+		}
 	}
 
 	taskOrPipelineName := huh.NewForm(
